pyproject.toml
[project]
name = "afterpython"
version = "0.1.0.dev17"
description = "CLI tool to streamline Python package maintenance and generate a dedicated project website"
license = "Apache-2.0"
authors = [
    { name = "Stephen Yau", email = "softwareentrepreneer+afterpython@gmail.com" },
]
readme = "README.md"
keywords = [
    "package-maintenance",
    "project-website",
    "static-site-generator",
    "developer-tools",
    "documentation",
    "cli-tool",
    "ci/cd",
    "mystmd",
    "semantic-release",
    "python-packaging",
]
classifiers = [
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Documentation",
    "Typing :: Typed",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    # "Programming Language :: Python :: 3.14",
]
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "trogon>=0.6.0",
    "httpx>=0.28.1",
    "tomlkit>=0.13.3",
    "ruamel-yaml>=0.18.16",
    "gitpython>=3.1.45",
    # "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
    "pyproject-metadata>=0.10.0",
    # "jupytext>=1.18.1",
    # "jinja2>=3.1.6",
    "mystmd>=1.7.0",
    "pdoc>=16.0.0",
    "ruff>=0.14.8",
    "pre-commit>=4.5.0",
    "commitizen>=4.10.0",
    # "python-semantic-release>=10.5.2",
]

[project.urls]
homepage = "https://afterpython.afterpython.org"
repository = "https://github.com/AfterPythonOrg/afterpython"
documentation = "https://afterpython.afterpython.org/doc"

[project.scripts]
ap = "afterpython.main:run_cli"
pcu = "afterpython.main:run_pcu"

[tool.uv]
# override the restrictive version constraints "platformdirs~=4.2.2", "nodeenv~=1.9.1" in package `mystmd`
override-dependencies = ["platformdirs>=4.5.0", "nodeenv>=1.9.1"]

[build-system]
requires = ["uv_build>=0.9.16,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "afterpython"
module-root = "src"
